Create test helpers module to provide basic data.
